playbook/roles/webapp/templates/webapp/requirements.txt
absl-py==0.4.1
astor==0.7.1
astroid==2.0.4
Click==7.0
cloudpickle==0.5.5
cycler==0.10.0
dask==0.19.1
decorator==4.3.0
Flask==1.0.2
Flask-Uploads==0.2.1
Flask-WTF==0.14.2
gast==0.2.0
grpcio==1.14.2
h5py==2.8.0
imgaug==0.2.6
isort==4.3.4
itsdangerous==0.24
Jinja2==2.10
kafka==1.3.5
Keras==2.2.2
Keras-Applications==1.0.4
Keras-Preprocessing==1.0.2
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==2.2.3
mccabe==0.6.1
networkx==2.1
numpy==1.14.5
opencv-python==3.4.3.18
pandas==0.23.4
Pillow==5.2.0
protobuf==3.6.1
pylint==2.1.1
pyparsing==2.2.0
python-dateutil==2.7.3
pytz==2018.5
PyWavelets==1.0.0
PyYAML==3.13
scikit-image==0.14.0
scikit-learn==0.19.2
scipy==1.1.0
six==1.11.0
sklearn==0.0
tensorboard==1.10.0
tensorflow-gpu==1.10.1
termcolor==1.1.0
toolz==0.9.0
typed-ast==1.1.0
Werkzeug==0.14.1
wrapt==1.10.11
WTForms==2.2.1
